+++ NEWS.md
@@ -8,6 +8,7 @@
   arguments (#5).
 - Make the `what` argument to `fuzz()` accept a list of inputs, so that
   the function can test multiple inputs in the same run.
+- Add a default list of inputs to `fuzz()`.
 
 # CBTF 0.2.0 (2025-07-16)